Handle delete tag,category, add user profiles, making administraion users

+++ CMS_blog/app/Http/Controllers/TagsController.php
@@ -98,6 +98,10 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        if($tag->posts->count()>0){
+            session()->flash('error','Tag can not deleted, because it is associated to some posts');
+            return redirect()->back();
+        }
         $tag->delete();
         session()->flash('success','Tag Deleted Successfully.');
         return  redirect(route('tags.index'));
